# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nom de votre API",
        "description": "Description de votre API",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Liste des utilisateurs",
                "description": "Renvoie la liste de tous les utilisateurs.",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs récupérée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Créer un nouvel utilisateur",
                "description": "Crée un nouvel utilisateur avec les détails fournis.",
                "operationId": "a1fce870c566eb5fc0cf3178ec5d2fed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom de l'utilisateur.",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Adresse e-mail de l'utilisateur.",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Mot de passe de l'utilisateur.",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur créé avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation des données en échec."
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Afficher les détails d'un utilisateur",
                "description": "Récupère les détails d'un utilisateur spécifié par son ID.",
                "operationId": "910a0c7cf580e1c605648d82a4e57f46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur à afficher.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'utilisateur récupérés avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé."
                    }
                }
            },
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Mettre à jour un utilisateur",
                "description": "Met à jour les informations d'un utilisateur spécifié par son ID.",
                "operationId": "ddfe232db1870b82493bf8bd25596d37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur à mettre à jour.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nouveau nom de l'utilisateur.",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Nouvelle adresse e-mail de l'utilisateur.",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Nouveau mot de passe de l'utilisateur.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé."
                    },
                    "422": {
                        "description": "Validation des données en échec."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Supprimer un utilisateur",
                "description": "Supprime un utilisateur spécifié par son ID.",
                "operationId": "43dd20becb1f47444e484cbb18ce345c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur à supprimer.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur supprimé avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé."
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Inscription d'un nouvel utilisateur",
                "description": "Inscrit un nouvel utilisateur avec les détails fournis.",
                "operationId": "f1fde7f4a932e393f6042d1b92e65b0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nom de l'utilisateur.",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Adresse e-mail de l'utilisateur.",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Mot de passe de l'utilisateur.",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmation du mot de passe de l'utilisateur.",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur inscrit avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "description": "Jetton d'accès JWT.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation des données en échec."
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Authentification de l'utilisateur",
                "description": "Authentifie l'utilisateur avec les informations d'identification fournies.",
                "operationId": "abc53c091ac89d7a2af0892997d1ee08",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Adresse e-mail de l'utilisateur.",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Mot de passe de l'utilisateur.",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur connecté avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Statut de la requête.",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur connecté."
                                        },
                                        "token": {
                                            "description": "Jetton d'accès JWT.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erreur d'authentification.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Statut de la requête.",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "L'e-mail ou le mot de passe est incorrect."
                                        },
                                        "errors": {
                                            "description": "Erreurs de validation.",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Statut de la requête.",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur interne du serveur."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "ID de l'utilisateur",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nom de l'utilisateur",
                        "type": "string"
                    },
                    "email": {
                        "description": "Adresse e-mail de l'utilisateur",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Mot de passe de l'utilisateur",
                        "type": "string",
                        "format": "password"
                    },
                    "created_at": {
                        "description": "Date de création de l'utilisateur",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour de l'utilisateur",
                        "type": "string",
                        "format": "date-time"
                    },
                    "profile_photo_url": {
                        "description": "URL de la photo de profil de l'utilisateur",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}